+++ checker_test.go
@@ -20,7 +20,7 @@ func (hc *HealthCheckerMock) Type() string {
 	return "Mock"
 }
 
-func (hc *HealthCheckerMock) Ping() error {
+func (hc *HealthCheckerMock) Check() error {
 	args := hc.Called()
 
 	return args.Error(0)
@@ -32,16 +32,16 @@ func TestStatusHealthChecker_Type(t *testing.T) {
 	assert.Equal(t, "Status", healthChecker.Type())
 }
 
-func TestStatusHealthChecker_Ping(t *testing.T) {
+func TestStatusHealthChecker_Check(t *testing.T) {
 	healthChecker := NewStatusHealthChecker(true)
 
-	assert.NoError(t, healthChecker.Ping())
+	assert.NoError(t, healthChecker.Check())
 }
 
-func TestStatusHealthChecker_Ping_Fail(t *testing.T) {
+func TestStatusHealthChecker_Check_Fail(t *testing.T) {
 	healthChecker := NewStatusHealthChecker(false)
 
-	err := healthChecker.Ping()
+	err := healthChecker.Check()
 
 	assert.Error(t, err)
 	assert.Equal(t, ErrHealthCheckFailed, err)
@@ -52,7 +52,7 @@ func TestStatusHealthChecker_SetStatus(t *testing.T) {
 
 	healthChecker.SetStatus(true)
 
-	assert.NoError(t, healthChecker.Ping())
+	assert.NoError(t, healthChecker.Check())
 }
 
 func TestDbHealthChecker_Type(t *testing.T) {
@@ -61,24 +61,24 @@ func TestDbHealthChecker_Type(t *testing.T) {
 	assert.Equal(t, "DatabasePing", healthChecker.Type())
 }
 
-// func TestDbHealthChecker_Ping(t *testing.T) {
+// func TestDbHealthChecker_Check(t *testing.T) {
 // 	db, _ := sql.Open("mysql", "obviously_wrong")
 
 // 	healthChecker := &DbHealthChecker{
 // 		db: db,
 // 	}
 
-// 	//assert.NoError(t, healthChecker.Ping())
+// 	//assert.NoError(t, healthChecker.Check())
 // }
 
-func TestDbHealthChecker_Ping_Fail(t *testing.T) {
+func TestDbHealthChecker_Check_Fail(t *testing.T) {
 	db, err := sql.Open("mysql", "user:password@/dbname")
 
 	require.NoError(t, err)
 
 	healthChecker := NewDbHealthChecker(db)
 
-	err = healthChecker.Ping()
+	err = healthChecker.Check()
 
 	assert.Error(t, err)
 }
@@ -89,7 +89,7 @@ func TestHTTPHealthChecker_Type(t *testing.T) {
 	assert.Equal(t, "HTTPPing", healthChecker.Type())
 }
 
-func TestHTTPHealthChecker_Ping(t *testing.T) {
+func TestHTTPHealthChecker_Check(t *testing.T) {
 	ts := httptest.NewServer(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
 		w.WriteHeader(http.StatusOK)
 		w.Write([]byte("ok"))
@@ -98,10 +98,10 @@ func TestHTTPHealthChecker_Ping(t *testing.T) {
 
 	healthChecker := NewHTTPHealthChecker(ts.URL)
 
-	assert.NoError(t, healthChecker.Ping())
+	assert.NoError(t, healthChecker.Check())
 }
 
-func TestHTTPHealthChecker_Ping_Fail(t *testing.T) {
+func TestHTTPHealthChecker_Check_Fail(t *testing.T) {
 	ts := httptest.NewServer(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
 		w.WriteHeader(http.StatusServiceUnavailable)
 		w.Write([]byte("error"))
@@ -110,7 +110,7 @@ func TestHTTPHealthChecker_Ping_Fail(t *testing.T) {
 
 	healthChecker := NewHTTPHealthChecker(ts.URL)
 
-	err := healthChecker.Ping()
+	err := healthChecker.Check()
 
 	assert.Error(t, err)
 	assert.Equal(t, ErrHealthCheckFailed, err)